tng_fixed/backend-python/requirements.txt
flask==3.1.0
flask-cors==5.0.1
google-generativeai==0.8.5
python-dotenv==1.0.1
gunicorn==23.0.0
